refactor category image handling to use Uri instead of Bitmap
